perf(ui): share Docker images data via React Context
Moved Docker images fetch from per-component hook to app-level
DockerImagesProvider. One fetch shared across all StartRunButtons,
workflow editors, and infrastructure page. 60s refresh interval
runs once regardless of how many components consume the data.
