+++ packages/platform-ui/src/app/(app)/layout.tsx
@@ -5,6 +5,7 @@ import { useRouter, usePathname } from 'next/navigation';
 import { useAuth } from '@/contexts/auth-context';
 import { AppShell } from '@/components/app-shell';
 import { CommandPaletteProvider } from '@/components/command-palette';
+import { DockerImagesProvider } from '@/contexts/docker-images-context';
 
 export default function AppLayout({ children }: { children: React.ReactNode }) {
   const { firebaseUser, loading, mustChangePassword } = useAuth();
@@ -33,8 +34,10 @@ export default function AppLayout({ children }: { children: React.ReactNode }) {
   if (!firebaseUser) return null;
 
   return (
-    <CommandPaletteProvider>
-      <AppShell>{children}</AppShell>
-    </CommandPaletteProvider>
+    <DockerImagesProvider>
+      <CommandPaletteProvider>
+        <AppShell>{children}</AppShell>
+      </CommandPaletteProvider>
+    </DockerImagesProvider>
   );
 }
