+++ packages/platform-ui/src/hooks/use-docker-images.ts
@@ -1,54 +1,8 @@
 'use client';
 
-import { useState, useEffect } from 'react';
-import { apiFetch } from '@/lib/api-fetch';
-import type { DockerImageInfo, DockerDiskInfo } from '@mediforce/platform-api/contract';
+export { useDockerImages } from '@/contexts/docker-images-context';
 
-interface DockerImagesState {
-  images: DockerImageInfo[];
-  disk: DockerDiskInfo | null;
-  isAvailable: boolean;
-  isLoading: boolean;
-}
-
-const REFRESH_INTERVAL_MS = 60_000;
-
-export function useDockerImages(): DockerImagesState {
-  const [state, setState] = useState<DockerImagesState>({
-    images: [],
-    disk: null,
-    isAvailable: false,
-    isLoading: true,
-  });
-
-  useEffect(() => {
-    let cancelled = false;
-
-    async function fetchDockerInfo() {
-      try {
-        const res = await apiFetch('/api/system/docker-info');
-        if (!res.ok || cancelled) return;
-        const data = await res.json();
-        if (cancelled) return;
-        if (data.available === true) {
-          setState({ images: data.images, disk: data.disk, isAvailable: true, isLoading: false });
-        } else {
-          setState({ images: [], disk: null, isAvailable: false, isLoading: false });
-        }
-      } catch {
-        if (!cancelled) {
-          setState({ images: [], disk: null, isAvailable: false, isLoading: false });
-        }
-      }
-    }
-
-    fetchDockerInfo();
-    const interval = setInterval(fetchDockerInfo, REFRESH_INTERVAL_MS);
-    return () => { cancelled = true; clearInterval(interval); };
-  }, []);
-
-  return state;
-}
+import type { DockerImageInfo } from '@mediforce/platform-api/contract';
 
 export function isImageAvailable(images: DockerImageInfo[], imageRef: string): boolean {
   const [repo, tag = 'latest'] = imageRef.split(':');
